Implementação da controller, serviços, correção de bugs, etc

+++ API/Controllers/RegisterController.cs
@@ -1,5 +1,6 @@
 ﻿using API.Data;
 using API.Entities;
+using API.Interface;
 using API.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -7,12 +8,13 @@ using Newtonsoft.Json;
 using RestSharp;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 
 namespace API.Controllers
 {
     [ApiController]
-    [Route("api/account")]
+    [Route("api/account/")]
     public class RegisterController : Controller
     {
         public RegisterController(LocalDb localDb, ZipCodeService zipCode)
@@ -20,24 +22,26 @@ namespace API.Controllers
             _localDb = localDb;
             _zipCode = zipCode;
             
-            
+
         }
         private LocalDb _localDb;
         private ZipCodeService _zipCode;
-             
-           
+        
+
+
 
         [HttpGet]
-        public IEnumerable<Client> GetClients()
+        public IEnumerable<IUser> GetClients()
         {
-            List<Client> clientsAddress = new List<Client>();
+            List<IUser> clientsAddress = new List<IUser>();
 
             var clients = _localDb.Clients.ToList();
             var address = _localDb.Address.ToList();
 
             foreach (var client in clients)
             {
-                Client client2 = client;
+                IUser client2 = new Client();
+                  client2  = client;
                 client2.Address = address.FirstOrDefault(a => a.AddressId == client.Tax_id);
                 clientsAddress.Add(client2);
             }
@@ -45,13 +49,60 @@ namespace API.Controllers
             return clientsAddress;
         }
 
-        
+       
+        [HttpGet("name/{name}")]
+        public IEnumerable<IUser> GetClientsFilterName(string name)
+        {
+            List<IUser> clientsAddress = new List<IUser>();
+
+            var clients = _localDb.Clients.ToList();
+            var address = _localDb.Address.ToList();
+
+             
+            foreach (var client in clients)
+            {
+                IUser client2 = new Client();
+                client2 = client;
+                client2.Address = address.FirstOrDefault(a => a.AddressId == client.Tax_id);
+                clientsAddress.Add(client2);
+            }
+
+            List<IUser> filterList = clientsAddress.Where(a => a.Name.ToLower() == name.ToLower()).ToList();
+
+            return filterList;
+        }
+
+        [HttpGet("date/")]
+        public IEnumerable<IUser> GetClientsFilterDate()
+        {
+            List<IUser> clientsAddress = new List<IUser>();
+
+            var clients = _localDb.Clients.ToList();
+            var address = _localDb.Address.ToList();
+
+
+            foreach (var client in clients)
+            {
+                IUser client2 = new Client();
+                client2 = client;
+                client2.Address = address.FirstOrDefault(a => a.AddressId == client.Tax_id);
+                clientsAddress.Add(client2);
+            }           
+           
+
+           List<IUser> filterList = clientsAddress.OrderByDescending(c => c.Created_at).ToList();
+
+            
+            return filterList;
+        }
+
 
         [HttpGet("{taxId}")]
         public IActionResult GetClientByTaxId(string taxId)
         {
-            
-                Client client = _localDb.Clients.FirstOrDefault(x => x.Tax_id == taxId);
+
+            IUser client = new Client();
+            client = _localDb.Clients.FirstOrDefault(x => x.Tax_id == taxId);
                 if(client == null)
             {
                 return NotFound("Usuario não encontrado!");
@@ -84,6 +135,7 @@ namespace API.Controllers
                 if (!client.Postal_Code.All(char.IsDigit) || string.IsNullOrWhiteSpace(client.Postal_Code))
             {
                 client.Status = "Pending";
+                client.Address.AddressId = client.Tax_id;
             }
             else
             {
@@ -106,9 +158,11 @@ namespace API.Controllers
         [HttpDelete("{taxId}")]
         public IActionResult DeleteClientByTaxId(string taxId)
         {
-            Client client = _localDb.Clients.FirstOrDefault(x => x.Tax_id == taxId);
-            Address address = _localDb.Address.FirstOrDefault(a => a.AddressId == taxId);
-            if (client != null)
+            IUser client = _localDb.Clients.FirstOrDefault(x => x.Tax_id == taxId);
+            IAddress address = _localDb.Address.FirstOrDefault(a => a.AddressId == taxId);
+
+            
+            if (client != null && client.Address != null)
             {
                 
                 _localDb.Remove(client);
